Extract TerminalPanel TextBox input subsystem to its own file
Sources/Panels/TerminalPanel.swift sat at 929 lines - 29 over the
900-line hard cap in scripts/swift_file_length_budget.py, with a
checked-in budget of exactly 929. Any PR that grows the file by even
one line now fails the "Validate Swift file length budget" CI step
with reason=hard cap 900 (e.g. task-notification-preserve-terminal-scroll
at actual=932 budget=929), which cascades into linux-preflight, tests,
and ci-status failures.

Move the TextBox composer subsystem (activation, focus arbitration,
escape handling, draft preservation/restoration, and the DEBUG inline
fixture) into TerminalPanel+TextBoxInput.swift, following the existing
TerminalPanel+NotificationScroll.swift / +SessionScrollbackReplay.swift
pattern. Stored TextBox state stays in the class (extensions cannot add
storage) and drops `private` so the extension file can reach it.

Pure mechanical move; no behavior changes. TerminalPanel.swift goes
from 929 to 573 lines and the new file is 365 lines (untracked by the
budget, threshold 500). No budget TSV change needed, so this cannot
conflict with other open PRs.
